BUG-1696 : Take new auditlog to use
All diffing of changes is now handled in the
auditlogger code. It works on Gson instead of Jackson,
which enforces some requirements on the Java POJOs to
be logged: they may not have fields with same name in
their inheritance hierarchies.

This is why BaseV1RDTO.oid and
NayttotutkintoV1RDTO.tutkintonimikes are now shared
with their subclasses, and the copy pasted equal
properties removed from the subclasses that had them.