Keep yhteystiedot if "use organization's address" is selected
Previous implementation was sending empty yhteystiedot array if
hakukohde was set to use its organization's hakuosoite. Koulutusinformaatio
module was not able to figure out the right organization's address because
it does not have the information about organizational hierarchy.

With this change, tarjonta will always send the hakuosoite, even if
it is a organization's osoite, so that koulutusinformatio has the
correct data to show.